Conditionally remove mfacp column from fire_alarm_control_panels

+++ db/migrate/20250108222910_remove_mfacp_from_fire_alarm_control_panels.rb
@@ -1,5 +1,9 @@
 class RemoveMfacpFromFireAlarmControlPanels < ActiveRecord::Migration[7.1]
   def change
-    remove_column :fire_alarm_control_panels, :mfacp, :string
+    if column_exists?(:fire_alarm_control_panels, :mfacp)
+      remove_column :fire_alarm_control_panels, :mfacp, :string
+    else
+      say "Column mfacp does not exist on fire_alarm_control_panels. Skipping removal."
+    end
   end
 end
